fixed constant

+++ src/main/java/frc/robot/Constants.java
@@ -92,6 +92,6 @@ public final class Constants {
     public static final double kInPosition = 0;
     public static final double kWristSpeed = 0.2;
 
-    public static final int kWristSpeedMultiplier = 0.3;
+    public static final double kWristSpeedMultiplier = 0.3;
   }
 }
